+++ requirements.txt
@@ -14,9 +14,4 @@ xarray>=0.20.1
 zarr>=2.10.0
 dask>=2022.0.0
 # for dev
-codespell>=2.0.0
-black>=21.9b0
-isort>=5.9.3
-flake8>=4.0.1
-pylint>=3.0.3
 pre-commit>=2.15.0
